Addressed PR comments

+++ clients/src/main/java/org/apache/kafka/common/protocol/Protocol.java
@@ -31,7 +31,6 @@ import static org.apache.kafka.common.protocol.types.Type.BYTES;
 import static org.apache.kafka.common.protocol.types.Type.INT16;
 import static org.apache.kafka.common.protocol.types.Type.INT32;
 import static org.apache.kafka.common.protocol.types.Type.INT64;
-import static org.apache.kafka.common.protocol.types.Type.INT8;
 import static org.apache.kafka.common.protocol.types.Type.LOG_BUFFER;
 import static org.apache.kafka.common.protocol.types.Type.STRING;
 import static org.apache.kafka.common.protocol.types.Type.NULLABLE_STRING;
@@ -945,7 +944,7 @@ public class Protocol {
                     INT16,
                     "The epoch for the PID. Will always be 0 if no appId was specified in the request."),
             new Field( "error_code",
-                    INT8,
+                    INT16,
                     "An integer error code.")
     );
